Refactoring Tan Verification (Abstract Factory) style fixes
