+++ src/main/java/app/coronawarn/verification/factories/tan/verifier/TanVerifierFactory.java
@@ -1,0 +1,7 @@
+package app.coronawarn.verification.factories.tan.verifier;
+
+import app.coronawarn.verification.model.AppSessionSourceOfTrust;
+
+public interface TanVerifierFactory {
+  public TanVerifier makeTanVerifier(AppSessionSourceOfTrust appSessionSourceOfTrust);
+}
