+++ src/main/java/app/coronawarn/verification/factories/test/identifier/CovidTestIdentifierFactory.java
@@ -18,20 +18,12 @@
  * ---license-end
  */
 
-package app.coronawarn.verification.factory;
+package app.coronawarn.verification.factories.test.identifier;
 
 import app.coronawarn.verification.model.RegistrationTokenRequest;
 
-public class CovidTestIdentifierFactoryImpl implements CovidTestIdentifierFactory {
-  @Override
-  public CovidTestIdentifier makeCovidTestIdentifier(RegistrationTokenRequest request) {
-    switch (request.getKeyType()) {
-      case GUID:
-        return new GuidCovidTestIdentifier();
-      case TELETAN:
-        return new TeleTanCovidTestIdentifier();
-      default:
-        return new UnknownCovidTestIdentifier();
-    }
-  }
+public interface CovidTestIdentifierFactory {
+
+  public CovidTestIdentifier makeCovidTestIdentifier(RegistrationTokenRequest request);
+
 }
