+++ src/main/java/app/coronawarn/verification/factories/test/identifier/TeleTanCovidTestIdentifier.java
@@ -18,24 +18,32 @@
  * ---license-end
  */
 
-package app.coronawarn.verification.factory;
+package app.coronawarn.verification.factories.test.identifier;
 
+import static java.util.concurrent.TimeUnit.MILLISECONDS;
+
+import app.coronawarn.verification.domain.VerificationTan;
+import app.coronawarn.verification.exception.VerificationServerException;
 import app.coronawarn.verification.model.RegistrationToken;
 import app.coronawarn.verification.model.RegistrationTokenRequest;
 import app.coronawarn.verification.service.AppSessionService;
 import app.coronawarn.verification.service.FakeDelayService;
 import app.coronawarn.verification.service.TanService;
+import java.util.Optional;
+import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.util.StopWatch;
 import org.springframework.web.context.request.async.DeferredResult;
 
 @RequiredArgsConstructor
 @Slf4j
-public class GuidCovidTestIdentifier extends CovidTestIdentifier {
+public class TeleTanCovidTestIdentifier extends CovidTestIdentifier {
 
+  private final ScheduledExecutorService scheduledExecutor = Executors.newScheduledThreadPool(4);
 
   @Override
   public DeferredResult<ResponseEntity<RegistrationToken>> generateRegistrationToken(
@@ -47,13 +55,26 @@ public class GuidCovidTestIdentifier extends CovidTestIdentifier {
     FakeDelayService fakeDelayService,
     TanService tanService) {
 
-    ResponseEntity<RegistrationToken> responseEntity =
-      appSessionService.generateRegistrationTokenByGuid(request.getKey(), request.getKeyDob(), fake);
+    Optional<VerificationTan> optional = tanService.getEntityByTan(request.getKey());
+
+    ResponseEntity<RegistrationToken> response = appSessionService.generateRegistrationTokenByTeleTan(
+      request.getKey(),
+      fake,
+      optional.map(VerificationTan::getTeleTanType).orElse(null));
+
+    if (optional.isPresent()) {
+      VerificationTan teleTan = optional.get();
+      teleTan.setRedeemed(true);
+      tanService.saveTan(teleTan);
+      stopWatch.stop();
+      fakeDelayService.updateFakeTokenRequestDelay(stopWatch.getTotalTimeMillis());
+      DeferredResult<ResponseEntity<RegistrationToken>> deferredResult = new DeferredResult<>();
+      scheduledExecutor.schedule(() -> deferredResult.setResult(response), fakeDelayService.realDelayToken(),
+        MILLISECONDS);
+      log.info("Returning the successfully generated RegistrationToken.");
+      return deferredResult;
+    }
     stopWatch.stop();
-    fakeDelayService.updateFakeTokenRequestDelay(stopWatch.getTotalTimeMillis());
-    DeferredResult<ResponseEntity<RegistrationToken>> deferredResult = new DeferredResult<>();
-    deferredResult.setResult(responseEntity);
-    log.info("Returning the successfully generated RegistrationToken.");
-    return deferredResult;
+    throw new VerificationServerException(HttpStatus.BAD_REQUEST, "The teleTAN verification failed");
   }
 }
